refactor(babel-plugin): rename files
