--- packages/babel-plugin-glaze/__test__/__fixtures__/defaultExport/code.js
@@ -1,15 +1,0 @@
-import React from 'react';
-
-export default () => {
-  return (
-    <p
-      sx={{
-        px: 4,
-        color: 'white',
-        bg: 'red',
-      }}
-    >
-      Hello, world!
-    </p>
-  );
-};
