--- packages/babel-plugin-glaze/__test__/__fixtures__/defaultExport/output.js
@@ -1,16 +1,0 @@
-import { useStyling } from 'glaze';
-import React from 'react';
-export default () => {
-  const sx = useStyling();
-  return (
-    <p
-      className={sx({
-        px: 4,
-        color: 'white',
-        bg: 'red',
-      })}
-    >
-      Hello, world!
-    </p>
-  );
-};
