--- packages/babel-plugin-glaze/__test__/__fixtures__/functionDeclaration/output.js
@@ -1,17 +1,0 @@
-import { useStyling } from 'glaze';
-import React from 'react';
-
-function App() {
-  const sx = useStyling();
-  return (
-    <p
-      className={sx({
-        px: 4,
-        color: 'white',
-        bg: 'red',
-      })}
-    >
-      Hello, world!
-    </p>
-  );
-}
