--- packages/babel-plugin-glaze/__test__/__fixtures__/hoc/code.js
@@ -1,17 +1,0 @@
-import React from 'react';
-
-const withTheme = (Component) => (props) => (
-  <Component sx={{ color: 'blue' }} {...props} />
-);
-
-const FancyButton = React.forwardRef((props, ref) => (
-  <button ref={ref} className="FancyButton" sx={{ color: 'blue' }}>
-    {props.children}
-  </button>
-));
-
-const MemoizedBUtton = React.memo((props) => (
-  <button ref={ref} className="MemoizedButton" sx={{ color: 'blue' }}>
-    {props.children}
-  </button>
-));
