--- packages/babel-plugin-glaze/__test__/__fixtures__/hoc/output.js
@@ -1,41 +1,0 @@
-import { useStyling } from 'glaze';
-import React from 'react';
-
-const withTheme = (Component) => (props) => {
-  const sx = useStyling();
-  return (
-    <Component
-      className={sx({
-        color: 'blue',
-      })}
-      {...props}
-    />
-  );
-};
-
-const FancyButton = React.forwardRef((props, ref) => {
-  const sx = useStyling();
-  return (
-    <button
-      ref={ref}
-      className={`FancyButton ${sx({
-        color: 'blue',
-      })}`}
-    >
-      {props.children}
-    </button>
-  );
-});
-const MemoizedBUtton = React.memo((props) => {
-  const sx = useStyling();
-  return (
-    <button
-      ref={ref}
-      className={`MemoizedButton ${sx({
-        color: 'blue',
-      })}`}
-    >
-      {props.children}
-    </button>
-  );
-});
