--- packages/babel-plugin-glaze/__test__/__fixtures__/hook/code.js
@@ -1,13 +1,0 @@
-import React from 'react';
-
-const WithHook = () => {
-  const useMyHook = () => {
-    const [state, useState] = React.useState('hello');
-
-    return <div sx={{ color: 'blue' }}>{state}</div>;
-  };
-
-  const toRender = useMyHook();
-
-  return <div>{toRender}</div>;
-};
