--- packages/babel-plugin-glaze/__test__/__fixtures__/importAlias/code.js
@@ -1,18 +1,0 @@
-import { useStyling as useCustomStyling } from 'glaze';
-import React from 'react';
-
-const App = () => {
-  const InnerComponent = () => <div sx={{ color: 'blue' }}>hello</div>;
-
-  return (
-    <p
-      sx={{
-        px: 4,
-        color: 'white',
-        bg: 'red',
-      }}
-    >
-      Hello, world!
-    </p>
-  );
-};
