--- packages/babel-plugin-glaze/__test__/__fixtures__/innerComponent/code.js
@@ -1,9 +1,0 @@
-import React from 'react';
-
-const Component = (props) => {
-  const InnerComponent = () => {
-    return <div sx={{ color: 'blue' }}>Inner</div>;
-  };
-
-  return <div sx={{ color: 'red' }}>{Component}</div>;
-};
