--- packages/babel-plugin-glaze/__test__/__fixtures__/innerComponent/output.js
@@ -1,29 +1,0 @@
-import { useStyling } from 'glaze';
-import React from 'react';
-
-const Component = (props) => {
-  const sx = useStyling();
-
-  const InnerComponent = () => {
-    const sx = useStyling();
-    return (
-      <div
-        className={sx({
-          color: 'blue',
-        })}
-      >
-        Inner
-      </div>
-    );
-  };
-
-  return (
-    <div
-      className={sx({
-        color: 'red',
-      })}
-    >
-      {Component}
-    </div>
-  );
-};
