--- packages/babel-plugin-glaze/__test__/__fixtures__/innerFunction/code.js
@@ -1,9 +1,0 @@
-import React from 'react';
-
-const Component = (props) => {
-  const renderComponent = () => {
-    return <div sx={{ color: 'blue' }}>Inner</div>;
-  };
-
-  return <div sx={{ hello: 'world' }}>{renderComponent()}</div>;
-};
