--- packages/babel-plugin-glaze/__test__/__fixtures__/innerFunction/output.js
@@ -1,28 +1,0 @@
-import { useStyling } from 'glaze';
-import React from 'react';
-
-const Component = (props) => {
-  const sx = useStyling();
-
-  const renderComponent = () => {
-    return (
-      <div
-        className={sx({
-          color: 'blue',
-        })}
-      >
-        Inner
-      </div>
-    );
-  };
-
-  return (
-    <div
-      className={sx({
-        hello: 'world',
-      })}
-    >
-      {renderComponent()}
-    </div>
-  );
-};
