--- packages/babel-plugin-glaze/__test__/__fixtures__/object/code.js
@@ -1,8 +1,0 @@
-import React from 'react';
-
-const createComp = () => ({
-  MyComp: (props) => {
-    return <div sx={{ color: 'blue' }}>Hello</div>;
-  },
-  Button: () => <button sx={{ color: 'red' }}>Button</button>,
-});
