--- packages/babel-plugin-glaze/__test__/__fixtures__/object/output.js
@@ -1,29 +1,0 @@
-import { useStyling } from 'glaze';
-import React from 'react';
-
-const createComp = () => ({
-  MyComp: (props) => {
-    const sx = useStyling();
-    return (
-      <div
-        className={sx({
-          color: 'blue',
-        })}
-      >
-        Hello
-      </div>
-    );
-  },
-  Button: () => {
-    const sx = useStyling();
-    return (
-      <button
-        className={sx({
-          color: 'red',
-        })}
-      >
-        Button
-      </button>
-    );
-  },
-});
