--- packages/babel-plugin-glaze/__test__/__fixtures__/somethingAlreadyImportedFromGlaze/code.js
@@ -1,16 +1,0 @@
-import React from 'react';
-import { something } from 'glaze';
-
-const App = () => {
-  return (
-    <p
-      sx={{
-        px: 4,
-        color: 'white',
-        bg: 'red',
-      }}
-    >
-      Hello, world!
-    </p>
-  );
-};
