--- packages/babel-plugin-glaze/__test__/__fixtures__/somethingAlreadyImportedFromGlaze/output.js
@@ -1,17 +1,0 @@
-import React from 'react';
-import { something, useStyling } from 'glaze';
-
-const App = () => {
-  const sx = useStyling();
-  return (
-    <p
-      className={sx({
-        px: 4,
-        color: 'white',
-        bg: 'red',
-      })}
-    >
-      Hello, world!
-    </p>
-  );
-};
