--- packages/babel-plugin-glaze/__test__/__fixtures__/sxExpression/code.js
@@ -1,13 +1,0 @@
-import React from 'react';
-
-const myStyle = { color: 'blue' };
-
-const myFn = () => ({ color: 'red' });
-
-const Comp1 = () => {
-  return <div sx={myStyle}></div>;
-};
-
-const Comp2 = () => {
-  return <div sx={myFn()}></div>;
-};
