--- packages/babel-plugin-glaze/__test__/__fixtures__/useStylingAlreadyImported/output.js
@@ -1,17 +1,0 @@
-import React from 'react';
-import { useStyling } from 'glaze';
-
-const App = () => {
-  const sx = useStyling();
-  return (
-    <p
-      className={sx({
-        px: 4,
-        color: 'white',
-        bg: 'red',
-      })}
-    >
-      Hello, world!
-    </p>
-  );
-};
