--- packages/babel-plugin-glaze/__test__/__fixtures__/withClassName/code.js
@@ -1,23 +1,0 @@
-import React from 'react';
-
-const SimpleClassName = (props) => {
-  return (
-    <div className="my-class-name" sx={{ color: 'blue' }}>
-      {Component}
-    </div>
-  );
-};
-
-const ComplexClassName = (props) => {
-  const fn = () => true;
-  return (
-    <div
-      className={`${props.className} other-class-name ${
-        fn('test') ? 'conditional' : ''
-      }`}
-      sx={{ color: 'blue' }}
-    >
-      {Component}
-    </div>
-  );
-};
