--- packages/babel-plugin-glaze/__test__/test.js
@@ -1,14 +1,0 @@
-import pluginTester from 'babel-plugin-tester';
-import path from 'path';
-
-import plugin from '../src/index';
-
-pluginTester({
-  plugin,
-  fixtures: path.join(__dirname, '__fixtures__'),
-  babelOptions: {
-    parserOpts: {
-      plugins: ['jsx'],
-    },
-  },
-});
