+++ packages/babel-plugin-glaze/tests/fixtures/default-export/code.jsx
@@ -1,0 +1,15 @@
+import * as React from 'react';
+
+export default () => {
+  return (
+    <p
+      sx={{
+        px: 4,
+        color: 'white',
+        bg: 'red',
+      }}
+    >
+      Hello, world!
+    </p>
+  );
+};
