+++ packages/babel-plugin-glaze/tests/fixtures/default-export/output.jsx
@@ -1,0 +1,16 @@
+import { useStyling } from 'glaze';
+import * as React from 'react';
+export default () => {
+  const sx = useStyling();
+  return (
+    <p
+      className={sx({
+        px: 4,
+        color: 'white',
+        bg: 'red',
+      })}
+    >
+      Hello, world!
+    </p>
+  );
+};
