+++ packages/babel-plugin-glaze/tests/fixtures/function-declaration/code.jsx
@@ -1,0 +1,15 @@
+import * as React from 'react';
+
+function App() {
+  return (
+    <p
+      sx={{
+        px: 4,
+        color: 'white',
+        bg: 'red',
+      }}
+    >
+      Hello, world!
+    </p>
+  );
+}
