+++ packages/babel-plugin-glaze/tests/fixtures/function-declaration/output.jsx
@@ -1,0 +1,17 @@
+import { useStyling } from 'glaze';
+import * as React from 'react';
+
+function App() {
+  const sx = useStyling();
+  return (
+    <p
+      className={sx({
+        px: 4,
+        color: 'white',
+        bg: 'red',
+      })}
+    >
+      Hello, world!
+    </p>
+  );
+}
