+++ packages/babel-plugin-glaze/tests/fixtures/function-expression/code.jsx
@@ -1,0 +1,15 @@
+import * as React from 'react';
+
+const App = function () {
+  return (
+    <p
+      sx={{
+        px: 4,
+        color: 'white',
+        bg: 'red',
+      }}
+    >
+      Hello, world!
+    </p>
+  );
+};
