+++ packages/babel-plugin-glaze/tests/fixtures/hook/output.jsx
@@ -1,0 +1,21 @@
+import { useStyling } from 'glaze';
+import * as React from 'react';
+
+const WithHook = () => {
+  const useMyHook = () => {
+    const sx = useStyling();
+    const [state, useState] = React.useState('hello');
+    return (
+      <div
+        className={sx({
+          color: 'blue',
+        })}
+      >
+        {state}
+      </div>
+    );
+  };
+
+  const toRender = useMyHook();
+  return <div>{toRender}</div>;
+};
