+++ packages/babel-plugin-glaze/tests/fixtures/inner-function/code.jsx
@@ -1,0 +1,9 @@
+import * as React from 'react';
+
+const Component = (props) => {
+  const renderComponent = () => {
+    return <div sx={{ color: 'blue' }}>Inner</div>;
+  };
+
+  return <div sx={{ hello: 'world' }}>{renderComponent()}</div>;
+};
