+++ packages/babel-plugin-glaze/tests/fixtures/object/code.jsx
@@ -1,0 +1,8 @@
+import * as React from 'react';
+
+const createComp = () => ({
+  MyComp: (props) => {
+    return <div sx={{ color: 'blue' }}>Hello</div>;
+  },
+  Button: () => <button sx={{ color: 'red' }}>Button</button>,
+});
