+++ packages/babel-plugin-glaze/tests/fixtures/object/output.jsx
@@ -1,0 +1,29 @@
+import { useStyling } from 'glaze';
+import * as React from 'react';
+
+const createComp = () => ({
+  MyComp: (props) => {
+    const sx = useStyling();
+    return (
+      <div
+        className={sx({
+          color: 'blue',
+        })}
+      >
+        Hello
+      </div>
+    );
+  },
+  Button: () => {
+    const sx = useStyling();
+    return (
+      <button
+        className={sx({
+          color: 'red',
+        })}
+      >
+        Button
+      </button>
+    );
+  },
+});
