+++ packages/babel-plugin-glaze/tests/fixtures/something-already-imported-from-glaze/code.jsx
@@ -1,0 +1,16 @@
+import { something } from 'glaze';
+import * as React from 'react';
+
+const App = () => {
+  return (
+    <p
+      sx={{
+        px: 4,
+        color: 'white',
+        bg: 'red',
+      }}
+    >
+      Hello, world!
+    </p>
+  );
+};
