+++ packages/babel-plugin-glaze/tests/fixtures/something-already-imported-from-glaze/output.jsx
@@ -1,0 +1,17 @@
+import { something, useStyling } from 'glaze';
+import * as React from 'react';
+
+const App = () => {
+  const sx = useStyling();
+  return (
+    <p
+      className={sx({
+        px: 4,
+        color: 'white',
+        bg: 'red',
+      })}
+    >
+      Hello, world!
+    </p>
+  );
+};
