+++ packages/babel-plugin-glaze/tests/fixtures/useStyling-already-imported/code.jsx
@@ -1,0 +1,16 @@
+import { useStyling } from 'glaze';
+import * as React from 'react';
+
+const App = () => {
+  return (
+    <p
+      sx={{
+        px: 4,
+        color: 'white',
+        bg: 'red',
+      }}
+    >
+      Hello, world!
+    </p>
+  );
+};
