+++ packages/babel-plugin-glaze/tests/fixtures/with-className/output.jsx
@@ -1,0 +1,33 @@
+import { useStyling } from 'glaze';
+import * as React from 'react';
+
+const SimpleClassName = (props) => {
+  const sx = useStyling();
+  return (
+    <div
+      className={`my-class-name ${sx({
+        color: 'blue',
+      })}`}
+    >
+      {Component}
+    </div>
+  );
+};
+
+const ComplexClassName = (props) => {
+  const sx = useStyling();
+
+  const fn = () => true;
+
+  return (
+    <div
+      className={`${`${props.className} other-class-name ${
+        fn('test') ? 'conditional' : ''
+      }`} ${sx({
+        color: 'blue',
+      })}`}
+    >
+      {Component}
+    </div>
+  );
+};
