+++ packages/babel-plugin-glaze/tests/index.test.js
@@ -1,0 +1,14 @@
+import pluginTester from 'babel-plugin-tester';
+import path from 'path';
+
+import plugin from '../src';
+
+pluginTester({
+  plugin,
+  fixtures: path.join(__dirname, 'fixtures/'),
+  babelOptions: {
+    parserOpts: {
+      plugins: ['jsx'],
+    },
+  },
+});
